New configs for the default report sorting: sortData.d.index, sortData.d.rev, sortData.m.index, sortData.m.rev, sortData.y.index, sortData.y.rev. Obsoleted and removed old global vars: sortColIndex, sortColRev, oldSortColIndex, oldSortColRev.
Using this new method, we store all the sort settings for each report. When changing reports, and coming back, the sorting will be the same as when you first left. No need for the ugly save/restore in changeReport() anymore.

This also helps the mission to unpollute the global namespace.