Convert line breaks from Windows (CR+LF) to Unix (LF)
All project files must be in Unix format.

Except when Windows Notepad.exe compatibility is desired, as in Browser
and Portable flavors, which are automatically generated and converted
to Windows.

From this point on, the Beta version files are not Notepad.exe friendly
anymore. Advanced users who opt for the Beta version could easily
install a better text editor.