Workaround lacking of SVN automatic revision number
In SVN one could use `$Revision$` to automatically expand to the
current revision number. Git doesn't have an equivalent for that, so we
must remove this feature.

Now the Cloud version will be the latest Git hash and will link to the
current commit (since it is generated, it's possible).

The Beta version will be infinite ∞ and will link to the repository
root.

Browser and Portable versions will continue to link to the current
version News page.