New storage driver: Google Drive
I have taken a simple approach for this first version: on the file
picker you select one or more text files and MoneyLog will open them.

For the future, selecting a folder would be best, so we can load all its
text files and even a possible config.js file.

Closes #8