toggleToolbarBox(): new option to force open/closed state
Before it was simpler because we relied on a known init state for all boxes:
closed.

Now that the Google Drive storage reads user config.js and reinit the UI, the
init state for the toolbar boxes is unknown. This new option forces a open or
closed state, respecting the user config.