Improve and fix user defaults for checkboxes
Before we were only acting if the user wanted to check the checkbox. If he wanted to uncheck it, no action was taken, and it worked because the initial state of all checkboxes is off.

But now Google Drive storage can reload the UI, and the current state for all checkboxes is not fixed.

With this change now we can force a check or uncheck.