googledrive: Refactor: load default folder, picker as fallback
Until now, the Picker was the only way to specify the MoneyLog folder
inside your Drive.

Now the the Picker is the fallback method. First we try to find a
MoneyLog folder in the root of your Drive. If found, its files are
loaded and we're done.

If not found, then we load the Picker API, and build the Picker so the
user can point where the MoneyLog folder is located.

See issue #25