Using oninput instead of onkeyup
The best event for detecting input changes is `input`. It triggers when the field changes by using the mouse (e.g. middle-clicking to paste, or through the context menu), and it doesn't trigger if key presses don't change the string.

There are other event handlers here (onclick?, onchange?) that might be worth revisiting and changing to a better event, but that's not going to happen in this commit (or ever).